Refactor: 관심지역, 높이 필터링 컴포넌트 각각 분리
